refactor: regenerate API files to fix problems with wrong return types
BREAKING CHANGE: To fix the problems, the return type of the POIKit.getRegionalPrice function had to be changed from RegionalPrices typealias to List<RegionalPrices>. The RegionalPrices object is now no longer a typealias, but a class that has the same properties as the old RegionalPrice model. The old RegionalPrice model no longer exists.