fix: catch ActivityNotFoundExeption if no supported browser is installed during authorization
BREAKING CHANGE: Introduced a NoSupportedBrowser exception which will now be returned from all IDKit.authorize(...) and IDKit.endSession(...) calls instead of crashing the app with an ActivityNotFoundException if no supported browser is installed and enabled to handle the Custom Tab intent.
In this case, we recommend displaying a dialog prompting the user to install a supported browser (e.g. Google Chrome or Firefox).
The suspendable calls IDKit.authorize(...) and IDKit.endSession(...) now return the Completion result instead of invoking a function parameter on result to make the calls synchronous.